% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddGeneBinScore.R
\name{AddGeneBinScore}
\alias{AddGeneBinScore}
\title{Add gene bin score}
\usage{
AddGeneBinScore(
  object,
  gene.trajectory,
  N.bin = 5,
  trajectories = 1:2,
  assay = "RNA",
  prefix = "Trajectory",
  reverse = NULL,
  echo = F
)
}
\arguments{
\item{object}{Seurat object}

\item{gene.trajectory}{Gene trajectory data frame}

\item{N.bin}{integer; How many gene bins}

\item{trajectories}{vector; Which gene trajectories to define gene bin score}

\item{assay}{string; Which assay to use}

\item{prefix}{string; String added to the names in the metadata}

\item{reverse}{vector; Whether to reverse the order of genes along each trajectory}

\item{echo}{boolean; Whether to print messages}
}
\value{
Seurat object with gene bin score added in the metadata
}
\description{
Add gene bin score
}
\examples{


}
