% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIgtPythonEnvironment.R
\name{checkIgtPythonEnvironment}
\alias{checkIgtPythonEnvironment}
\title{Check the Python environment for running IGT (Inference of Gene Trajectories) using reticulate package in R}
\usage{
checkIgtPythonEnvironment(
  python.packages = c("ot", "tqdm", "scipy.io"),
  check.numpy = TRUE
)
}
\description{
Check the Python environment for running IGT (Inference of Gene Trajectories) using reticulate package in R
}
