% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarse.grain.R
\name{coarse.grain}
\alias{coarse.grain}
\title{Base function called by CoarseGrain}
\usage{
coarse.grain(
  cell.embedding,
  gene.expression,
  graph.dist,
  N = 1000,
  random.seed = 1
)
}
\arguments{
\item{cell.embedding}{matrix; Cell embedding for kNN clustering}

\item{gene.expression}{matrix; Cell-by-gene count matrix (library normalized)}

\item{graph.dist}{matrix; Cell-cell graph distance matrix}

\item{N}{integer; Number of meta-cells after coarse-graining}

\item{random.seed;}{Random seed for kNN clustering}
}
\value{
Returns a list
}
\description{
Base function called by CoarseGrain
}
\examples{


}
