% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoarseGrain.R
\name{CoarseGrain}
\alias{CoarseGrain}
\title{Coarse-graining}
\usage{
CoarseGrain(
  object,
  graph.dist,
  features,
  N = 1000,
  assay = "RNA",
  reduction = "dm",
  dims = 1:5,
  slot = "data",
  random.seed = 1
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{vector;}

\item{N}{integer;}

\item{assay}{string;}

\item{reduction}{string;}

\item{dims}{vector;}

\item{slot}{string;}

\item{random.seed}{integer;}
}
\value{
Returns a list
}
\description{
Coarse-graining
}
