% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.map.R
\name{diffusion.map}
\alias{diffusion.map}
\title{Run Diffusion Map on a precomputed distance matrix}
\usage{
diffusion.map(dist.mat, K = 10, sigma = NULL, nEV = 30, t = 1)
}
\arguments{
\item{dist.mat}{matrix; Precomputed distance matrix (symmetric)}

\item{K}{integer; Adaptive kernel bandwidth for each point set to be the distance to its \code{K}-th nearest neighbor.}

\item{sigma}{numeric; Fixed kernel bandwidth, \code{sigma} will be ignored if \code{K} is specified.}

\item{nEV}{integer; Number of leading eigenvectors to export}

\item{t}{integer; Number of diffusion times}
}
\value{
List with the following elements:
\tabular{ll}{
\code{diffu.emb} \tab Diffusion Map embedding with the leading \code{K} eigenvectors \cr
\tab \cr
\code{eigen.vals} \tab Eigenvalues associated with corresponding eigenvectors \cr
}
}
\description{
Run Diffusion Map on a precomputed distance matrix
}
\examples{


}
