% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractGeneTrajectory.R
\name{ExtractGeneTrajectory}
\alias{ExtractGeneTrajectory}
\title{Extract gene trajectories}
\usage{
ExtractGeneTrajectory(
  gene.embedding,
  dist.mat,
  N,
  t.list,
  dims = 1:5,
  K = 10,
  quantile = 0.02
)
}
\arguments{
\item{gene.embedding}{matrix; Gene embedding}

\item{dist.mat}{matrix; Gene-gene Wasserstein distance matrix (symmetric)}

\item{N}{integer; Number of gene trajectories to retrieve}

\item{t.list}{integer; Number of diffusion times to retrieve each trajectory}

\item{dims}{vector; Dimensions of gene embedding to use to identify terminal genes (extrema)}

\item{K}{integer; Adaptive kernel bandwidth for each point set to be the distance to its \code{K}-th nearest neighbor.}

\item{quantile}{numeric; Thresholding parameter to extract genes for each trajectory. Default: 0.02}
}
\value{
A data frame indicating gene trajectories and gene ordering along each trajectory
}
\description{
Extract gene trajectories
}
\examples{


}
