% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.RW.matrix.R
\name{get.RW.matrix}
\alias{get.RW.matrix}
\title{Convert a distance matrix into a random-walk matrix based on adaptive Gaussian kernel}
\usage{
get.RW.matrix(dist.mat, K = 10)
}
\arguments{
\item{dist.mat}{matrix; Precomputed distance matrix (symmetric)}

\item{K}{integer; Adaptive kernel bandwidth for each point set to be the distance to its \code{K}-th nearest neighbor.}
}
\value{
Random-walk matrix
}
\description{
Convert a distance matrix into a random-walk matrix based on adaptive Gaussian kernel
}
\examples{


}
