% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGeneEmbedding.R
\name{GetGeneEmbedding}
\alias{GetGeneEmbedding}
\title{Get the diffusion embedding of genes based on the gene-gene Wasserstein distance matrix}
\usage{
GetGeneEmbedding(dist.mat, K = 10, sigma = NULL, nEV = 30, t = 1)
}
\arguments{
\item{dist.mat}{matrix; Gene-gene Wasserstein distance matrix (symmetric)}

\item{K}{integer; Adaptive kernel bandwidth for each point set to be the distance to its \code{K}-th nearest neighbor.}

\item{sigma}{numeric; Fixed kernel bandwidth, \code{sigma} will be ignored if \code{K} is specified.}

\item{nEV}{integer; Number of leading eigenvectors to export}

\item{t}{integer; Number of diffusion times}
}
\value{
List with the following elements:
\tabular{ll}{
\code{diffu.emb} \tab Diffusion Map embedding with the leading \code{K} eigenvectors \cr
\tab \cr
\code{eigen.vals} \tab Eigenvalues associated with corresponding eigenvectors \cr
}
}
\description{
Get the diffusion embedding of genes based on the gene-gene Wasserstein distance matrix
}
