% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGenePseudoorder.R
\name{GetGenePseudoorder}
\alias{GetGenePseudoorder}
\title{Extract gene trajectories}
\usage{
GetGenePseudoorder(dist.mat, subset, max.id = NULL)
}
\arguments{
\item{dist.mat}{matrix; Gene-gene Wasserstein distance matrix (symmetric)}

\item{subset}{vector; Genes in a given trajectory}

\item{max.id}{integer; Index of the terminal gene}
}
\value{
Pseudoorder of genes along a given trajectory
}
\description{
Extract gene trajectories
}
\examples{


}
