% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGraphDistance.R
\name{GetGraphDistance}
\alias{GetGraphDistance}
\title{Run Diffusion Map on a Seurat object}
\usage{
GetGraphDistance(object, reduction = "dm", dims = 1:5, K = 10)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{string; Dimensionality reduction to use, default: 'dm'}

\item{dims}{integer; Which dimensions to use as input features for kNN graph construction}

\item{K}{integer; Adaptive kernel bandwidth for each point set to be the distance to its \code{K}-th nearest neighbor.}
}
\value{
Returns a cell-cell graph distance matrix
}
\description{
Run Diffusion Map on a Seurat object
}
\examples{


}
