% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadGeneDistMat.R
\name{LoadGeneDistMat}
\alias{LoadGeneDistMat}
\title{Load gene-gene Wasserstein matrix}
\usage{
LoadGeneDistMat(dir.path, file_name = "emd.csv")
}
\arguments{
\item{dir.path}{Path to gene-gene Wasserstein matrix}

\item{file.name}{Name of the matrix file}
}
\value{
Gene-gene Wasserstein matrix
}
\description{
Load gene-gene Wasserstein matrix
}
