% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupIgtPythonEnvironment.R
\name{setupIgtPythonEnvironment}
\alias{setupIgtPythonEnvironment}
\title{Set up Python environment for running IGT (Inference of Gene Trajectories) using reticulate package in R}
\usage{
setupIgtPythonEnvironment(
  condaenv = "igt",
  virtualenv = "igt",
  packages = c("numpy", "scipy", "pot", "tqdm"),
  create = FALSE
)
}
\description{
Set up Python environment for running IGT (Inference of Gene Trajectories) using reticulate package in R
}
