% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.bifurcation.R
\name{simulate.bifurcation}
\alias{simulate.bifurcation}
\title{Simulate bifurcating gene process}
\usage{
\method{simulate}{bifurcation}(
  N_cells = 5 * c(100, 50, 50),
  N_genes = 5 * c(100, 50, 50),
  model = "poisson",
  meanlog = 0,
  sdlog = 0.25,
  scale = 25,
  seed = 1,
  maxT = 15,
  sort = TRUE,
  sparsity = 0.1,
  theta = 10
)
}
\arguments{
\item{N_cells}{integer vector; Number of cells in the parent process and two daughter processes}

\item{N_genes}{integer vector; Number of genes associated with the parent process and two daughter processes}

\item{model}{character; Count model ("poisson" or "negbin")}

\item{meanlog}{numeric; Mean of log normal distribution}

\item{sdlog}{numeric; Standard deviation of log normal distribution}

\item{scale}{numeric; Scale of UMI counts}

\item{seed}{integer; Random seed}

\item{maxT}{numeric; Maximum cell pseudotime}

\item{sort}{boolean; Whether to sort genes based on their peak times}

\item{sparsity}{numeric; Sparsity of count matrix}

\item{theta}{numeric; Dipersion parameter for negative binomial model}
}
\value{
Returns a gene-by-cell count matrix
}
\description{
Simulate bifurcating gene process
}
