% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.coral.R
\name{simulate.coral}
\alias{simulate.coral}
\title{Simulate a multi-level lineage differentiation process coupled with cell cycle}
\usage{
\method{simulate}{coral}(
  N_cells = 15 * rep(100, 7),
  N_genes = 2 * rep(100, 7),
  N_genes_CC = 400,
  model = "poisson",
  meanlog = 0,
  sdlog = 0.25,
  scale = 25,
  scale_l = 2,
  seed = 1,
  maxT = 10,
  maxT_CC = 20,
  sparsity = NULL,
  theta = 10
)
}
\arguments{
\item{N_cells}{integer vector; Number of cells associated with the differentiation process}

\item{N_genes}{integer vector; Number of genes associated with the cell cycle process}

\item{model}{character; Count model ("poisson" or "negbin")}

\item{meanlog}{numeric; Mean of log normal distribution}

\item{sdlog}{numeric; Standard deviation of log normal distribution}

\item{scale}{numeric; Scale of UMI counts}

\item{seed}{integer; Random seed}

\item{maxT}{numeric; Maximum cell pseudotime of each lineage in the differentiation process}

\item{maxT_CC}{numeric; Maximum cell pseudotime of the cell cycle process}

\item{sparsity}{numeric; Sparsity of count matrix}

\item{theta}{numeric; Dipersion parameter for negative binomial model}
}
\value{
Returns a gene-by-cell count matrix
}
\description{
Simulate a multi-level lineage differentiation process coupled with cell cycle
}
