% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.cylinder.R
\name{simulate.cylinder}
\alias{simulate.cylinder}
\title{Simulate concurrent independent linear and cyclic gene processes}
\usage{
\method{simulate}{cylinder}(
  N_cells = 5000,
  N_genes = rep(500, 2),
  model = "poisson",
  meanlog = 0,
  sdlog = 0.25,
  scale = 25,
  seed = 1,
  maxT = 15,
  sort = TRUE,
  sparsity = 0.1,
  theta = 10
)
}
\arguments{
\item{N_cells}{integer; Number of cells}

\item{N_genes}{integer vector; Number of genes associated with the linear and cyclic processes}

\item{model}{character; Count model ("poisson" or "negbin")}

\item{meanlog}{numeric; Mean of log normal distribution}

\item{sdlog}{numeric; Standard deviation of log normal distribution}

\item{scale}{numeric; Scale of UMI counts}

\item{seed}{integer; Random seed}

\item{maxT}{numeric; Maximum cell pseudotime}

\item{sort}{boolean; Whether to sort genes based on their peak times}

\item{sparsity}{numeric; Sparsity of count matrix}

\item{theta}{numeric; Dipersion parameter for negative binomial model}
}
\value{
Returns a gene-by-cell count matrix
}
\description{
Simulate concurrent independent linear and cyclic gene processes
}
