% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.linear.R
\name{simulate.linear}
\alias{simulate.linear}
\title{Simulate linear gene process}
\usage{
\method{simulate}{linear}(
  N_cells = 1000,
  N_genes = 200,
  model = "poisson",
  meanlog = 0,
  sdlog = 0.25,
  scale = 25,
  seed = 1,
  maxT = 15,
  sort = TRUE,
  sparsity = NULL,
  theta = 10
)
}
\arguments{
\item{N_cells}{integer; Number of cells}

\item{N_genes}{integer; Number of genes}

\item{model}{character; Count model ("poisson" or "negbin")}

\item{meanlog}{numeric; Mean of log normal distribution}

\item{sdlog}{numeric; Standard deviation of log normal distribution}

\item{scale}{numeric; Scale of UMI counts}

\item{seed}{integer; Random seed}

\item{maxT}{numeric; Maximum cell pseudotime}

\item{sort}{boolean; Whether to sort genes based on their peak times}

\item{sparsity}{numeric; Sparsity of count matrix}

\item{theta}{numeric; Dipersion parameter for negative binomial model}
}
\value{
Returns a gene-by-cell count matrix
}
\description{
Simulate linear gene process
}
